GH-90 Extract helper methods for process execution
